+++ mobile/android/android-components/buildSrc/src/main/java/Dependencies.kt
@@ -33,8 +33,6 @@ object Versions {
     const val disklrucache = "2.0.2"
     const val leakcanary = "2.8.1"
 
-    // When upgrading mozilla_appservices, also upgrade the version in the
-    // `getApplicationServiceVersion()` method in NimbusGradlePlugin.groovy.
     const val mozilla_appservices = "94.2.0"
 
     // DO NOT MODIFY MANUALLY. This is auto-updated along with GeckoView.