+++ mobile/android/android-components/docs/changelog.md
@@ -32,6 +32,9 @@ permalink: /changelog/
   * ⚠️ **This is a breaking change**: `FxaPushSupportFeature` now requires to be explicitly started with `initialize`.
   * The constructor for `FxaPushSupportFeature` has a `coroutineScope` parameter that defaults to a `CoroutineScope(Dispatchers.IO)`.
 
+* **nimbus-gradle-plugin**:
+  * Updated the plugin to use the version of application services defined in the buildSrc Dependencies. 
+
 # 105.0.0
 * [Commits](https://github.com/mozilla-mobile/android-components/compare/v104.0.0...v105.0.0)
 * [Milestone](https://github.com/mozilla-mobile/android-components/milestone/152?closed=1)